requirements.txt
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
flatbuffers==23.3.3
fonttools==4.39.3
humanfriendly==10.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.7.1
mltu==1.0.8
mpmath==1.3.0
numba==0.56.4
numpy==1.23.5
onnx==1.12.0
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.0
pooch==1.7.0
protobuf==3.20.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
resampy==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
sympy==1.11.1
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
